fold hop-engine-beam into hop-engines-beam, fixes #7235
