+++ assemblies/client/pom.xml
@@ -63,12 +63,6 @@
         </dependency>
         <dependency>
             <groupId>org.apache.hop</groupId>
-            <artifactId>hop-engine-beam</artifactId>
-            <version>${project.version}</version>
-            <type>zip</type>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.hop</groupId>
             <artifactId>hop-engines-beam</artifactId>
             <version>${project.version}</version>
             <type>zip</type>
