+++ plugins/engines/beam/src/main/java/org/apache/hop/beam/core/BeamDefaults.java
