+++ plugins/engines/beam/src/main/java/org/apache/hop/beam/core/HopRow.java
