+++ plugins/transforms/memgroupby/pom.xml
@@ -47,11 +47,5 @@
             <artifactId>commons-math3</artifactId>
             <scope>provided</scope>
         </dependency>
-        <dependency>
-            <groupId>org.apache.hop</groupId>
-            <artifactId>hop-engine-beam</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
     </dependencies>
 </project>
